+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+
+  - Service image state distributor and action added
+  - Component image state distributor and action added
   - Deployment log fetch action added
 
 ### Changed